receiver/requirements.txt
fastapi[all]
uvicorn[standard]
python-multipart
argon2-cffi
uuid
